% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{drayleigh}
\alias{drayleigh}
\title{Rayleigh's Probability Density Function}
\usage{
drayleigh(x, sigma)
}
\arguments{
\item{x}{Value to return its density}

\item{sigma}{The variance associated}
}
\value{
The probability density function of \code{x}
}
\description{
Function to calculate the Rayleigh probability density function
}
