% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mdc_rcpp}
\alias{mdc_rcpp}
\title{Maximo Divisor Comum}
\usage{
mdc_rcpp(x, y)
}
\arguments{
\item{x}{Um número (int)}

\item{y}{Um número (int)}
}
\value{
O MDC entre \code{x} e \code{y}
}
\description{
Maximo Divisor Comum
}
